Enabled more clang-tidy checks (#448)
* Enabled more clang-tidy checks

* Enabled more clang-tidy checks

+++ Source/Term/source/ScreenBuffer.cpp
@@ -2,6 +2,7 @@
 #include "ScreenBuffer.h"
 #include <CoreFoundation/CoreFoundation.h>
 #include <algorithm>
+#include <cstddef>
 
 namespace nc::term {
 
@@ -498,7 +499,7 @@ ScreenBuffer::Snapshot::Snapshot() : width(0), height(0)
 }
 
 ScreenBuffer::Snapshot::Snapshot(unsigned _w, unsigned _h)
-    : width(_w), height(_h), chars(std::make_unique<Space[]>(_w * _h))
+    : width(_w), height(_h), chars(std::make_unique<Space[]>(static_cast<size_t>(_w) * static_cast<size_t>(_h)))
 {
 }
 
@@ -517,9 +518,9 @@ void ScreenBuffer::RevertToSnapshot(const Snapshot &_snapshot)
     else { // TODO: anchor?
         std::fill_n(m_OnScreenSpaces.get(), m_Width * m_Height, m_EraseChar);
         for( int y = 0, e = std::min(_snapshot.height, m_Height); y != e; ++y ) {
-            std::copy_n(_snapshot.chars.get() + (y * _snapshot.width),
+            std::copy_n(_snapshot.chars.get() + (static_cast<size_t>(y * _snapshot.width)),
                         std::min(_snapshot.width, m_Width),
-                        m_OnScreenSpaces.get() + (y * m_Width));
+                        m_OnScreenSpaces.get() + (static_cast<size_t>(y * m_Width)));
         }
     }
 }
